+++ jvm-agent/src/main/org/jetbrains/kotlinx/lincheck/transformation/LincheckClassFileTransformer.kt
@@ -10,7 +10,7 @@
 
 package org.jetbrains.kotlinx.lincheck.transformation
 
-import org.jetbrains.kotlinx.lincheck.SMAPInfo
+import org.jetbrains.kotlinx.lincheck.transformation.SMAPInfo
 import org.jetbrains.kotlinx.lincheck.transformation.InstrumentationMode.*
 import org.jetbrains.kotlinx.lincheck.transformation.LincheckJavaAgent.INSTRUMENT_ALL_CLASSES
 import org.jetbrains.kotlinx.lincheck.transformation.LincheckJavaAgent.instrumentationMode
