+++ jvm-agent/src/main/org/jetbrains/kotlinx/lincheck/transformation/SMAPInfo.kt
@@ -8,9 +8,9 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.lincheck
+package org.jetbrains.kotlinx.lincheck.transformation
 
-import java.util.*
+import java.util.Objects
 
 /*
   This code was heavily adapted from `SDE.java` from IDEA monorepo.
